Revert "5"
This reverts commit 05d8d5d3724421c485e018bec37fa710443b643e.

+++ springboot/src/main/java/com/demo/controller/TeacherController.java
@@ -35,7 +35,7 @@ public class TeacherController {
 
     @GetMapping("{id}")
     @CrossOrigin("*")
-    public Teacher getById(@PathVariable int id) {
+    public Teacher getById(@PathVariable Long id) {
 
         return teacherService.find(id);
     }
@@ -48,7 +48,7 @@ public class TeacherController {
 
     @DeleteMapping("{id}")
     @CrossOrigin("*")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable Long id) {
         teacherService.delete(id);
     }
